requirements.txt
requests
python-whois
termcolor
dnspython
shodan
beautifulsoup4
lxml